requirements.txt
langchain>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
openai>=1.10.0
transformers>=4.36.0
sentence-transformers>=2.2.2
torch>=2.0.0
chromadb>=0.4.15
beautifulsoup4>=4.12.0
requests>=2.31.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
tqdm>=4.66.0
prompt-toolkit>=3.0.0
colorama>=0.4.6